// typpings.d.ts
import { Image } from "sanity";

type Base = {
  _id: string;
  _rev: string;
  _type: string;
  _updatedAt: string;
  _createdAt: string;
};

interface Post extends Base {
  title: string;
  slug: Slug;
  desciption: string;
  mainImage: Image;
  categories: Category[];
  body: Block[];
  author: Author;
}

interface Author extends Base {
  bio: Block[];
  image: Image;
  name: string;
  slug: Slug;
}

interface Reference {
  _ref: string;
  _type: "reference";
}

interface Slug {
  _type: "slug";
  current: string;
}

interface Block {
  _key: string;
  _type: "block";
  children: Span[];
  markDefs: any[];
  style: "normal" | "h1" | "h2" | "h3" | "h4" | "blockquote";
}

interface Span {
  _key: string;
  _type: "span";
  marks: string[];
  text: string;
}

interface Category extends Base {
  title: string;
  description: string;
}

interface MainImage {
  _type: "image";
  asset: Reference;
}
interface Title {
  _type: "title";
  current: string;
}